# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Land Registry API",
        "description": "Land register lookup API. Study project: the data is fictitious and does not describe real properties.",
        "license": {
            "name": "MIT"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "/",
            "description": "This server"
        }
    ],
    "paths": {
        "/api/v1/orders": {
            "post": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterOrder"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/titles": {
            "get": {
                "tags": [
                    "title-controller"
                ],
                "operationId": "search",
                "parameters": [
                    {
                        "name": "postcode",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 1,
                            "pattern": "^\\s*[A-Za-z]{1,2}[0-9][A-Za-z0-9]?\\s*[0-9][A-Za-z]{2}\\s*$"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20,
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TitleSearchPage"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/titles/{titleNumber}": {
            "get": {
                "tags": [
                    "title-controller"
                ],
                "operationId": "get",
                "parameters": [
                    {
                        "name": "titleNumber",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TitleDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{reference}": {
            "get": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "get_1",
                "parameters": [
                    {
                        "name": "reference",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterOrder"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateOrderRequest": {
                "type": "object",
                "properties": {
                    "titleNumber": {
                        "type": "string",
                        "minLength": 1
                    },
                    "documentType": {
                        "type": "string",
                        "enum": [
                            "TITLE_REGISTER",
                            "TITLE_PLAN"
                        ]
                    },
                    "applicantName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "applicantEmail": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "applicantAddress": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "applicantAddress",
                    "applicantEmail",
                    "applicantName",
                    "documentType",
                    "titleNumber"
                ]
            },
            "RegisterOrder": {
                "type": "object",
                "properties": {
                    "reference": {
                        "type": "string"
                    },
                    "titleNumber": {
                        "type": "string"
                    },
                    "documentType": {
                        "type": "string",
                        "enum": [
                            "TITLE_REGISTER",
                            "TITLE_PLAN"
                        ]
                    },
                    "applicantName": {
                        "type": "string"
                    },
                    "applicantEmail": {
                        "type": "string"
                    },
                    "applicantAddress": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING_PAYMENT",
                            "PAID"
                        ]
                    },
                    "amountPence": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Address": {
                "type": "object",
                "properties": {
                    "line1": {
                        "type": "string"
                    },
                    "line2": {
                        "type": "string"
                    },
                    "town": {
                        "type": "string"
                    },
                    "postcode": {
                        "type": "string"
                    }
                }
            },
            "TitleSearchPage": {
                "type": "object",
                "properties": {
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TitleSummary"
                        }
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "previousCursor": {
                        "type": "string"
                    },
                    "nextCursor": {
                        "type": "string"
                    }
                }
            },
            "TitleSummary": {
                "type": "object",
                "properties": {
                    "titleNumber": {
                        "type": "string"
                    },
                    "tenure": {
                        "type": "string",
                        "enum": [
                            "FREEHOLD",
                            "LEASEHOLD"
                        ]
                    },
                    "address": {
                        "$ref": "#/components/schemas/Address"
                    }
                }
            },
            "Charge": {
                "type": "object",
                "properties": {
                    "lender": {
                        "type": "string"
                    },
                    "chargeDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "amountPence": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "Proprietor": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    }
                }
            },
            "TitleDetail": {
                "type": "object",
                "properties": {
                    "titleNumber": {
                        "type": "string"
                    },
                    "tenure": {
                        "type": "string",
                        "enum": [
                            "FREEHOLD",
                            "LEASEHOLD"
                        ]
                    },
                    "classOfTitle": {
                        "type": "string"
                    },
                    "address": {
                        "$ref": "#/components/schemas/Address"
                    },
                    "pricePaidPence": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "pricePaidDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "proprietors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Proprietor"
                        }
                    },
                    "charges": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Charge"
                        }
                    }
                }
            }
        }
    }
}